Restructured code. Split types into different files, renamed files.
