+++ stats/tagging/key.go
@@ -1,0 +1,33 @@
+// Copyright 2017 Google Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+package tagging
+
+// Key is the interface for all key types.
+type Key interface {
+	Name() string
+}
+
+// KeyType defines the types of keys allowed.
+type keyType byte
+
+const (
+	// keyTypeUnknown is not a valid KeyType. It is here just to detect that a
+	// keyType isn't set.
+	keyTypeUnknown keyType = iota
+	keyTypeString
+	keyTypeBool
+	keyTypeInt64
+)
