+++ stats/tagging/key_string.go
@@ -1,0 +1,48 @@
+// Copyright 2017 Google Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+package tagging
+
+import "fmt"
+
+// keyString is implementation for keys which values are of type string.
+type keyString struct {
+	name string
+}
+
+func (ks *keyString) Name() string {
+	return ks.name
+}
+
+func (ks *keyString) CreateMutation(v string, mb MutationBehavior) *mutationString {
+	return &mutationString{
+		tagString: &tagString{
+			keyString: ks,
+			v:         v,
+		},
+		behavior: mb,
+	}
+}
+
+func (ks *keyString) createTag(s string) *tagString {
+	return &tagString{
+		keyString: ks,
+		v:         s,
+	}
+}
+
+func (ks *keyString) String() string {
+	return fmt.Sprintf("%T:'%s'", ks, ks.name)
+}
